Main Menu/Settings
Created Main Menu with Following Functions:

-Play
-Options
-Quit

Added UI sound effects for buttons & test music for main menu music.

Settings menu is fully functional and can adjust volume of master, music, sfx with sliders.

I made a AudioPlayer script thats global in order to play audio SFX and Music between scenes. However one bug I have not fixed on it is the main menu music will start at the beginning when you press back on the settings menu. This is because the main menu and settings menu are seperate scenes in godot. Will need to fix this later.

I changed some of the project settings so the screen size is dynamic for different resolutions and sizes of monitors. This can be altered if its an issue.